Check for if MEGA or VSCodium is already installed.

+++ Linux-Mint_19.3.sh
@@ -19,13 +19,23 @@ echo Installing packages.....
     curl -sS https://download.spotify.com/debian/pubkey.gpg | sudo apt-key add - 
     echo "deb http://repository.spotify.com stable non-free" | sudo tee /etc/apt/sources.list.d/spotify.list
 
-    wget https://mega.nz/linux/MEGAsync/xUbuntu_18.04/amd64/megasync-xUbuntu_18.04_amd64.deb -P /tmp/
-    dpkg -i /tmp/megasync-xUbuntu_18.04_amd64.deb
-    rm /tmp/megasync-xUbuntu_18.04_amd64.deb
+    if [[ $(dpkg --list | grep -wo megasync) = "megasync" ]]
+    then
+        echo Mega Sync is already installed...
+    else
+        wget https://mega.nz/linux/MEGAsync/xUbuntu_18.04/amd64/megasync-xUbuntu_18.04_amd64.deb -P /tmp/
+        dpkg -i /tmp/megasync-xUbuntu_18.04_amd64.deb
+        rm /tmp/megasync-xUbuntu_18.04_amd64.deb
+    fi
 
-    wget https://github.com/VSCodium/vscodium/releases/download/1.41.1/codium_1.41.1-1576787344_amd64.deb -P /tmp/
-    dpkg -i /tmp/codium_1.41.1-1576787344_amd64.deb
-    rm /tmp codium_1.41.1-1576787344_amd64.deb
+    if [[ $(dpkg --list | grep -wo codium) = "codium" ]]
+    then
+        echo VSCodium is already installed...
+    else
+        wget https://github.com/VSCodium/vscodium/releases/download/1.41.1/codium_1.41.1-1576787344_amd64.deb -P /tmp/
+        dpkg -i /tmp/codium_1.41.1-1576787344_amd64.deb
+        rm /tmp codium_1.41.1-1576787344_amd64.deb
+    fi
 
     apt-get update -y
     apt-get install -f -y
